backend/requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
google-genai
pydantic>=2.0.0
python-dotenv>=1.0.0
PyPDF2>=3.0.0
python-docx>=1.0.0
python-multipart>=0.0.9
numpy>=1.26.0
